refactor ui tabs, new media file
use new ui button sprites of the highlighted button and text
individually, while drawing the background blue box and black line in
software. this allows us to use the same cache for compact mode
(previously the longer inventory required a separate longer sprite), UIs
can be moved around more easily in the future, and sprite quality is
improved.

media58 is still supported and accounted for.